HSEARCH-3102 Move backend configurations for tests from documentation module to dedicated utils modules
So that we can use them in other modules
