+++ documentation/src/test/java/org/hibernate/search/documentation/backend/lucene/analyzer/LuceneGetAnalyzerIT.java
@@ -14,7 +14,7 @@ import javax.persistence.EntityManagerFactory;
 import org.hibernate.search.backend.lucene.LuceneBackend;
 import org.hibernate.search.backend.lucene.index.LuceneIndexManager;
 import org.hibernate.search.documentation.testsupport.DocumentationSetupHelper;
-import org.hibernate.search.documentation.testsupport.LuceneBackendConfiguration;
+import org.hibernate.search.documentation.testsupport.DocumentationLuceneBackendConfiguration;
 import org.hibernate.search.engine.backend.Backend;
 import org.hibernate.search.engine.backend.index.IndexManager;
 import org.hibernate.search.mapper.orm.Search;
@@ -30,7 +30,7 @@ public class LuceneGetAnalyzerIT {
 
 	@Rule
 	public DocumentationSetupHelper setupHelper =
-			DocumentationSetupHelper.withSingleBackend( new LuceneBackendConfiguration() );
+			DocumentationSetupHelper.withSingleBackend( new DocumentationLuceneBackendConfiguration() );
 
 	private EntityManagerFactory entityManagerFactory;
