--- documentation/src/test/java/org/hibernate/search/documentation/testsupport/AbstractDocumentationBackendConfiguration.java
@@ -1,29 +1,0 @@
-/*
- * Hibernate Search, full-text search for your domain model
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.search.documentation.testsupport;
-
-import java.util.Map;
-
-import org.hibernate.search.util.impl.integrationtest.common.TestConfigurationProvider;
-import org.hibernate.search.util.impl.integrationtest.common.rule.BackendConfiguration;
-import org.hibernate.search.util.impl.integrationtest.common.rule.MappingSetupHelper;
-
-abstract class AbstractDocumentationBackendConfiguration implements BackendConfiguration {
-
-	@Override
-	public <C extends MappingSetupHelper<C, ?, ?>.AbstractSetupContext> C setup(C setupContext,
-			String backendNameOrNull, TestConfigurationProvider configurationProvider) {
-		return setupContext
-				.withBackendProperties(
-						backendNameOrNull,
-						configurationProvider.interpolateProperties( getBackendProperties() )
-				);
-	}
-
-	protected abstract Map<String, Object> getBackendProperties();
-
-}
