+++ documentation/src/test/java/org/hibernate/search/documentation/testsupport/DocumentationElasticsearchBackendConfiguration.java
@@ -1,0 +1,24 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.documentation.testsupport;
+
+import java.util.Map;
+
+import org.hibernate.search.util.impl.integrationtest.backend.elasticsearch.ElasticsearchBackendConfiguration;
+
+public class DocumentationElasticsearchBackendConfiguration extends ElasticsearchBackendConfiguration {
+
+	@Override
+	protected Map<String, Object> getBackendProperties() {
+		Map<String, Object> properties = super.getBackendProperties();
+		properties.put(
+				"analysis.configurer",
+				new ElasticsearchSimpleMappingAnalysisConfigurer()
+		);
+		return properties;
+	}
+}
